+++ PostHog/PostHogConfig.swift
@@ -13,27 +13,6 @@ import Foundation
         case any
     }
 
-    /// Determines the behavior for processing user profiles.
-    /// - `never`: We won't process persons for any event. This means that anonymous users will not be merged once they sign up or login, so you lose the ability to create funnels that track users from anonymous to identified. All events (including `$identify`) will be sent with `$process_person_profile: False`.
-    /// - `always`: We will process persons data for all events.
-    /// - `identifiedOnly`: (default): we will only process persons when you call `identify`, `alias`, and `group`, Anonymous users won't get person profiles.
-    @objc(PostHogPersonProfiles) public enum PostHogPersonProfiles: Int {
-        case never
-        case always
-        case identifiedOnly
-
-        var description: String {
-            switch self {
-            case .never:
-                return "never"
-            case .always:
-                return "always"
-            case .identifiedOnly:
-                return "identified_only"
-            }
-        }
-    }
-
     @objc public let host: URL
     @objc public let apiKey: String
     @objc public var flushAt: Int = 20