+++ PostHogTests/PostHogSDKTest.swift
@@ -17,7 +17,8 @@ class PostHogSDKTest: QuickSpec {
                 captureApplicationLifecycleEvents: Bool = false,
                 flushAt: Int = 1,
                 optOut: Bool = false,
-                propertiesSanitizer: PostHogPropertiesSanitizer? = nil) -> PostHogSDK
+                propertiesSanitizer: PostHogPropertiesSanitizer? = nil,
+                personProfiles: PostHogPersonProfiles = .identifiedOnly) -> PostHogSDK
     {
         let config = PostHogConfig(apiKey: "123", host: "http://localhost:9001")
         config.flushAt = flushAt
@@ -28,6 +29,7 @@ class PostHogSDKTest: QuickSpec {
         config.captureApplicationLifecycleEvents = captureApplicationLifecycleEvents
         config.optOut = optOut
         config.propertiesSanitizer = propertiesSanitizer
+        config.personProfiles = personProfiles
         return PostHogSDK.with(config)
     }
 
@@ -791,6 +793,73 @@ class PostHogSDKTest: QuickSpec {
 
             sut.close()
         }
+
+        it("capture sets process person to false if identified only and not identified") {
+            let sut = self.getSut()
+
+            sut.capture("test event",
+                        properties: ["foo": "bar"],
+                        userProperties: ["userProp": "value"],
+                        userPropertiesSetOnce: ["userPropOnce": "value"],
+                        groups: ["groupProp": "value"])
+
+            let events = getBatchedEvents(server)
+
+            expect(events.count) == 1
+
+            let event = events.first!
+
+            expect(event.properties["$process_person_profile"] as? Bool) == false
+
+            sut.reset()
+            sut.close()
+        }
+
+        it("capture sets process person to true if identified only and identified") {
+            let sut = self.getSut(flushAt: 2)
+
+            sut.identify("distinctId",
+                         userProperties: ["userProp": "value"],
+                         userPropertiesSetOnce: ["userPropOnce": "value"])
+
+            sut.capture("test event",
+                        properties: ["foo": "bar"],
+                        userProperties: ["userProp": "value"],
+                        userPropertiesSetOnce: ["userPropOnce": "value"],
+                        groups: ["groupProp": "value"])
+
+            let events = getBatchedEvents(server)
+
+            expect(events.count) == 2
+
+            let event = events.last!
+
+            expect(event.properties["$process_person_profile"] as? Bool) == true
+
+            sut.reset()
+            sut.close()
+        }
+
+//        it("capture sets process person to true if always") {
+//            let sut = self.getSut(personProfiles: .always)
+//
+//            sut.capture("test event",
+//                        properties: ["foo": "bar"],
+//                        userProperties: ["userProp": "value"],
+//                        userPropertiesSetOnce: ["userPropOnce": "value"],
+//                        groups: ["groupProp": "value"])
+//
+//            let events = getBatchedEvents(server)
+//
+//            expect(events.count) == 1
+//
+//            let event = events.first!
+//
+//            expect(event.properties["$process_person_profile"] as? Bool) == true
+//
+//            sut.reset()
+//            sut.close()
+//        }
     }
 }