+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 - recording: mask swiftui picker if masking enabled ([#184](https://github.com/PostHog/posthog-ios/pull/184))
 - chore: add is identified property ([#186](https://github.com/PostHog/posthog-ios/pull/186))
-- chore: add `personProfiles` support ([#186](https://github.com/PostHog/posthog-ios/pull/186))
+- chore: add personProfiles support ([#187](https://github.com/PostHog/posthog-ios/pull/187))
 
 ## 3.9.1 - 2024-09-06